Bashrc: bash_completion sourcing
Different systems put bash_completion in different places, and some will
source the file from /etc/profile if it exists, so we need a way to
configure this.

If needed, define $BASH_COMPLETION in .bashrc.local to contain the path
to bash_completion. .bashrc will look for the variable and, if the
location it points to is readable, will source it. Note that
$BASH_COMPLETION is removed from the environment after sourcing as it is
presumably not needed anymore.