Add my-init script with terminfo customizations
The script will contain any run-once sort of initializations I might
need when I'm standing up an account on a new machine.

The terminfo customizations pertain primarily to tmux, where I set
'flash' (the visual bell entry) to be the same as 'bel' (the audio bell
entry). I do this because I try to use a visual bell whenever possible,
and Tmux doesn't seem to meaningfully respond to it. This customization
makes programs send the audio bell, which I've configured within Tmux to
respond to with a visual message.