WM-independent re-write
Lots of changes to support working in multiple window managers
environments. The biggest is the introduction of the 'my' script, which
supports multiple utility functions which I'll then map to keys. The
keys are mapped using Xbindkeys to maintain consistency and avoid the
work of having to map them separately in each WM.