Refactor photosync.sh into cloudsync.sh
What was being used for syncing photos only when I was using a cloud
host is now syncing "mobile" documents (i.e., any files in
~/Documents/Mobile).

Also make more efficient by only syncing photos in the folder
representing the current month. Additionally look at a lastrun file to
see if the previous month needs to be synced as well.

One limitation of this solution is that it expects the sync will never
be more than a month behind, which is to say that we only sync the month
shown in 'lastrun' if it differs from the current month and *NOT* all
intervening months.