Move vim-css-color to optional packages
Not sure whether this is a result of recent changes to the vim-css-color
plugin itself, Vim, or my recent move to Arch Linux, but any filetype
that triggers the vim-css-color plugin logic upon opening a new buffer
makes startup *extremely* slow (multiple seconds). Moving the plugin to
the opt path makes it possible to load the plugin on demand, thus
eliminating the problem unless I explicitly decide to use it.