profile, shrc: Always print window title
The attempt to print a window title had, up until this point, only
occurred in profile, which means it only happened at login. This
approach does not make sense, since there is no guarantee that
profile is sourced in an interactive environment. This should
instead be done in shrc, when an actual, interactive shell is
started.

This commit makes that change.